Add option to run tests on local selenium driver

+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import base64
 import os
 import pytest
+import shutil
 import subprocess
 
 from pathlib import Path
@@ -8,18 +9,58 @@ from PIL import Image
 from selenium import webdriver
 
 
+def pytest_addoption(parser):
+    parser.addoption(
+        "--webdriver",
+        help="Selenium webdriver selection",
+        default="remote",
+    )
+    parser.addoption(
+        "--remote-executor",
+        help="Address of command executor for remote webdriver",
+        default="http://localhost:4444/wd/hub",
+    )
+    parser.addoption(
+        "--website-selenium",
+        help="Website address for selenium host",
+        default="http://app:8080",
+    )
+    parser.addoption(
+        "--backend-test-host",
+        help="Backed address for test host",
+        default="http://localhost:8080",
+    )
+
+
 def is_circleci():
     return "CI" in os.environ
 
 
 @pytest.fixture(scope="session")
-def selenium_data_path():
-    return "/home/seluser/data"
+def driver_option(request):
+    driver = request.config.getoption("--webdriver")
+    if driver != "remote" and driver != "firefox":
+        raise ValueError(f"Unsupported --webdriver value: {driver}")
+    return driver
 
 
 @pytest.fixture(scope="session")
-def website():
-    return "http://app:8080"
+def selenium_data_path(driver_option):
+    if driver_option == "remote":
+        # in container path:
+        return "/home/seluser/data"
+    return "/tmp/selenium"
+
+
+@pytest.fixture(scope="session")
+def website_selenium(request):
+    return request.config.getoption("--website-selenium")
+
+
+@pytest.fixture(scope="session")
+def pcb_endpoint(request):
+    value = request.config.getoption("--backend-test-host")
+    return f"{value}/api/pcb"
 
 
 def to_base64(path):
@@ -48,15 +89,19 @@ def pytest_runtest_makereport(item, call):
 
 
 @pytest.fixture(scope="session")
-def driver(selenium_data_path):
+def driver(request, driver_option, selenium_data_path):
     options = webdriver.FirefoxOptions()
     options.set_preference("browser.download.folderList", 2)
     options.set_preference("browser.download.manager.showWhenStarting", False)
     options.set_preference("browser.download.useDownloadDir", True)
     options.set_preference("browser.download.dir", f"{selenium_data_path}/downloads")
-    _driver = webdriver.Remote(
-        command_executor="http://localhost:4444/wd/hub", options=options
-    )
+    if driver_option == "remote":
+        _driver = webdriver.Remote(
+            command_executor=request.config.getoption("--remote-executor"), options=options
+        )
+        _driver.set_window_size(1360, 1800)
+    else:
+        _driver = webdriver.Firefox(options=options)
     yield _driver
     _driver.quit()
 
@@ -97,8 +142,8 @@ def full_screenshot(driver, tmpdir):
 
 
 @pytest.fixture
-def selenium(driver, website, tmpdir):
-    driver.get(website)
+def selenium(driver, website_selenium, tmpdir):
+    driver.get(website_selenium)
     yield driver
     full_screenshot(driver, tmpdir)
 
@@ -111,23 +156,29 @@ def run_command_in_container(container_id, command):
 
 
 @pytest.fixture
-def download_dir(request, selenium_data_path):
-    outputs_path = f"{request.config.rootdir}/data/downloads"
-    os.makedirs(outputs_path, exist_ok=True)
-    # mode change needed for selenium to write downloads there:
-    os.chmod(outputs_path, 0o777)
+def download_dir(request, driver_option, selenium_data_path):
+    # returns download directory from host perspective (if runninng in remote container)
+    if driver_option == "remote":
+        outputs_path = f"{request.config.rootdir}/data/downloads"
+        os.makedirs(outputs_path, exist_ok=True)
+        # mode change needed for selenium to write downloads there:
+        os.chmod(outputs_path, 0o777)
 
-    yield outputs_path
+        yield outputs_path
 
-    # cleanup requires workarounds...
-    # if selenium is run inside container, then downloaded file file has different file owner
-    # than process running tests. The workaround is to run rm in running container.
-    container_details = subprocess.run(
-        "docker container ls --all | grep 'selenium/standalone-firefox.*tests_firefox_[0-9]\+'",
-        shell=True, capture_output=True
-    )
-    if container_details.returncode == 0:
-        container_id = container_details.stdout.decode().split(" ", 1)[0]
-        run_command_in_container(container_id, f"rm -rf {selenium_data_path}/downloads")
+        # cleanup requires workarounds...
+        # if selenium is run inside container, then downloaded file file has different file owner
+        # than process running tests. The workaround is to run rm in running container.
+        container_details = subprocess.run(
+            "docker container ls --all | grep 'selenium/standalone-firefox.*tests_firefox_[0-9]\+'",
+            shell=True, capture_output=True
+        )
+        if container_details.returncode == 0:
+            container_id = container_details.stdout.decode().split(" ", 1)[0]
+            run_command_in_container(container_id, f"rm -rf {selenium_data_path}/downloads")
+        else:
+            raise Exception("Could not clean up downloads")
     else:
+        outputs_path = f"{selenium_data_path}/downloads"
+        yield outputs_path
         shutil.rmtree(outputs_path)
